refactor: create cli module
