+++ src/ape/cli/options.py
@@ -31,3 +31,6 @@ def verbose_option(help=""):
         default=False,
         help=help,
     )
+
+
+__all__ = ["network_option", "verbose_option"]
