+++ src/ape/managers/compilers.py
@@ -4,9 +4,10 @@ from typing import Dict, List, Set
 from dataclassy import dataclass
 
 from ape.api.compiler import CompilerAPI
+from ape.cli.utils import notify
 from ape.plugins import PluginManager
 from ape.types import ContractType
-from ape.utils import cached_property, notify
+from ape.utils import cached_property
 
 from ..exceptions import CompilerError
 from .config import ConfigManager
