+++ src/ape_run/_cli.py
@@ -5,8 +5,9 @@ from pathlib import Path
 import click
 
 from ape import config, networks
-from ape.options import network_option, verbose_option
-from ape.utils import Abort, get_relative_path
+from ape.cli.options import network_option, verbose_option
+from ape.cli.utils import Abort
+from ape.utils import get_relative_path
 from ape_console._cli import console
 
 # TODO: Migrate this to a CLI toolkit under ``ape``
